VEnv

Documentation for VEnv.

VEnv.addenv

VEnv.create

VEnv.getenv

VEnv.run

VEnv.addenv — Method

addenv(cmd::Cmd, venv::AbstractString; inherit=true)

Add environment variables to cmd so that it will run in the virtual environment at venv.

VEnv.create — Method

create([f], venv)

Create a Python Virtual Environment at the directory venv.

If f is given, then f(venv) is also called. It should be used to install packages into the environment.

Keyword Arguments

system_site_packages=false

symlinks=false

copies=false

clear=true: Remove any existing packages from the environment.

upgrade=false

without_pip=false

prompt=nothing

upgrade_deps=false

version=nothing: If this is given, it must be a string identifying the version of the environment. If the environment has already been created and has the same version, then creating the environment is skipped.

force=false: Do no skip creating the environment, even if the versions match.

VEnv.getenv — Method

getenv(venv; inherit=true)

A dictionary of environment variables in which the given venv is activated.

VEnv.run — Method

run(cmd, venv; inherit=true)

Runs the command cmd in the virtual environment at venv.