VEnv

Documentation for VEnv.

VEnv.addenv

VEnv.create

VEnv.create_once

VEnv.getenv

VEnv.run

VEnv.addenv — Method

addenv(cmd::Cmd, venv::AbstractString; inherit=true)

Add environment variables to cmd so that it will run in the virtual environment at venv.

If inherit is true then the current environment ENV is included too.

For example, the following prints out the version of Python in the venv at ./venv:

run(VEnv.addenv(`python --version`, "./venv"))

See also python and pip.

VEnv.create — Method

create(venv)

Create a Python Virtual Environment at the directory venv.

VEnv.create_once — Method

create_once([f], venv, version; kw...)

Create a virtual environment if it has not already been created.

Calls create(venv; kw...) if the virtual environment at venv does not exist or if the version has changed since the last time it was created.

You may optionally also specify a function f(venv) which is called when creating the environment.

VEnv.getenv — Method

getenv(venv; inherit=true)

A dictionary of environment variables in which the given venv is activated.

If inherit is true (the default) then the current ENV is included.

VEnv.run — Method

run(cmd, venv; inherit=true)

Runs the command cmd in the virtual environment at venv.

Simply shorthand for Base.run(VEnv.addenv(...)).